backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
websockets==14.2
python-dotenv==1.1.0
pydantic==2.11.1
pydantic-settings==2.9.1
langchain==0.3.25
langchain-openai==0.3.18
langchain-anthropic==0.3.15
langchain-community==0.3.24
httpx==0.28.1
gitpython==3.1.44
PyGithub==2.6.0
aiosqlite==0.21.0
aiofiles==24.1.0
